+++ tests/test_on_data/ophys/test_segmentation_interfaces.py
@@ -12,6 +12,7 @@ from neuroconv.datainterfaces import (
     CnmfeSegmentationInterface,
     ExtractSegmentationInterface,
     InscopixSegmentationInterface,
+    MinianSegmentationInterface,
     Suite2pSegmentationInterface,
 )
 from neuroconv.tools.testing.data_interface_mixins import (
@@ -606,3 +607,46 @@ class TestInscopixSegmentationInterfaceCellSetPart1(SegmentationExtractorInterfa
         assert "raw" in metadata["Ophys"]["Fluorescence"]["PlaneSegmentation"]
         raw_traces_metadata = metadata["Ophys"]["Fluorescence"]["PlaneSegmentation"]["raw"]
         assert raw_traces_metadata["name"] == "RoiResponseSeries"
+
+
+class TestMinianSegmentationInterface(SegmentationExtractorInterfaceTestMixin):
+    data_interface_cls = MinianSegmentationInterface
+    interface_kwargs = dict(
+        folder_path=OPHYS_DATA_PATH / "segmentation_datasets" / "minian" / "segmented_data_3units_100frames"
+    )
+    save_directory = OUTPUT_PATH
+
+    @pytest.fixture(
+        params=[
+            {"mask_type": "image", "include_background_segmentation": True},
+            {"mask_type": "pixel", "include_background_segmentation": True},
+            {"include_roi_centroids": False, "include_background_segmentation": True},
+            {"include_roi_acceptance": False, "include_background_segmentation": True},
+            {"include_background_segmentation": False},
+        ],
+        ids=[
+            "mask_type_image",
+            "mask_type_pixel",
+            "exclude_roi_centroids",
+            "exclude_roi_acceptance",
+            "exclude_background_segmentation",
+        ],
+    )
+    def setup_interface(self, request):
+
+        test_id = request.node.callspec.id
+        self.test_name = test_id
+        self.interface_kwargs = self.interface_kwargs
+        self.conversion_options = request.param
+        self.interface = self.data_interface_cls(**self.interface_kwargs)
+
+        return self.interface, self.test_name
+
+
+class TestMinianSegmentationInterfaceWithStubTest(SegmentationExtractorInterfaceTestMixin):
+    data_interface_cls = MinianSegmentationInterface
+    interface_kwargs = dict(
+        folder_path=OPHYS_DATA_PATH / "segmentation_datasets" / "minian" / "segmented_data_3units_100frames",
+    )
+    save_directory = OUTPUT_PATH
+    conversion_options = dict(stub_test=True)